tests/test_requirements.txt
anaconda-auth
anaconda-anon-usage
conda-anaconda-telemetry

httpx
anyio
certifi
httpcore==1.*
idna
h11
sniffio
typing_extensions